# docs/tenant-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "summary": "Register new user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Login user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "summary": "Refresh access token",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "summary": "Logout user",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/auth/mfa/enroll": {
      "post": {
        "operationId": "AuthController_enrollMfa",
        "summary": "Start MFA enrollment",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/auth/mfa/enable": {
      "post": {
        "operationId": "AuthController_enableMfa",
        "summary": "Enable MFA",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaCodeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/auth/mfa/disable": {
      "post": {
        "operationId": "AuthController_disableMfa",
        "summary": "Disable MFA",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaCodeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/auth/accept-invite": {
      "post": {
        "operationId": "AuthController_acceptInvite",
        "summary": "Accept tenant invitation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInviteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "AuthController_getMe",
        "summary": "Get current user",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/activities": {
      "get": {
        "operationId": "ActivitiesController_getAllActivities",
        "summary": "Get all activities with optional filters and pagination",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-indexed)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search query string",
            "schema": {
              "example": "revenue",
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Sort field and direction (e.g., createdAt:desc, name:asc)",
            "schema": {
              "example": "createdAt:desc",
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Filter by category (Leads, Sales, Operations, etc.)",
            "schema": {
              "example": "Sales",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by status",
            "schema": {
              "example": "Active",
              "enum": [
                "Active",
                "Completed",
                "Cancelled"
              ],
              "type": "string"
            }
          },
          {
            "name": "priority",
            "required": false,
            "in": "query",
            "description": "Filter by priority",
            "schema": {
              "example": "High",
              "enum": [
                "Low",
                "Medium",
                "High"
              ],
              "type": "string"
            }
          },
          {
            "name": "dueDateFrom",
            "required": false,
            "in": "query",
            "description": "Filter activities due on or after this date (ISO 8601)",
            "schema": {
              "example": "2025-01-01",
              "type": "string"
            }
          },
          {
            "name": "dueDateTo",
            "required": false,
            "in": "query",
            "description": "Filter activities due on or before this date (ISO 8601)",
            "schema": {
              "example": "2025-12-31",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns paginated activities assigned to the authenticated tenant user"
          },
          "400": {
            "description": "Bad Request – validation failed for payload or query parameters"
          },
          "401": {
            "description": "Unauthorized – missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden – insufficient role or tenant scope"
          }
        },
        "tags": [
          "Activities"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "post": {
        "operationId": "ActivitiesController_createActivity",
        "summary": "Create activity",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateActivityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Activity created for current tenant"
          },
          "400": {
            "description": "Bad Request – validation failed for payload or query parameters"
          },
          "401": {
            "description": "Unauthorized – missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden – insufficient role or tenant scope"
          }
        },
        "tags": [
          "Activities"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/activities/{id}": {
      "put": {
        "operationId": "ActivitiesController_updateActivity",
        "summary": "Update activity",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateActivityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated activity details"
          },
          "400": {
            "description": "Bad Request – validation failed for payload or query parameters"
          },
          "401": {
            "description": "Unauthorized – missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden – insufficient role or tenant scope"
          }
        },
        "tags": [
          "Activities"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "ActivitiesController_deleteActivity",
        "summary": "Delete activity",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Activity deleted"
          },
          "400": {
            "description": "Bad Request – validation failed for payload or query parameters"
          },
          "401": {
            "description": "Unauthorized – missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden – insufficient role or tenant scope"
          }
        },
        "tags": [
          "Activities"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/activities/summary": {
      "get": {
        "operationId": "ActivitiesController_getSummary",
        "summary": "Get activities summary for dashboard",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Aggregated activity summary"
          },
          "400": {
            "description": "Bad Request – validation failed for payload or query parameters"
          },
          "401": {
            "description": "Unauthorized – missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden – insufficient role or tenant scope"
          }
        },
        "tags": [
          "Activities"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "BG Accountability Tenant API",
    "description": "REST API for the BG Accountability Platform - Tenant Services.\n\n## Overview\nThis API serves both the Tenant Web Dashboard (Next.js) and Mobile App (Flutter).\n\n## Authentication\nAll protected endpoints require a valid JWT Bearer token.\n- Login via `POST /api/v1/auth/login`\n- Include token in header: `Authorization: Bearer <token>`\n- Refresh tokens via `POST /api/v1/auth/refresh`\n\n## Multi-Factor Authentication (MFA)\n- If MFA is enabled, login returns a `tempToken`\n- Complete MFA via `POST /api/v1/auth/mfa/login` with tempToken + TOTP code\n\n## Rate Limiting\n- Default: 100 requests per minute per IP\n- Auth endpoints: 10 requests per minute per IP\n\n## Error Format\nAll errors follow: `{ success: false, message: string, code?: string, details?: object }`\n\n## Pagination\nPaginated endpoints use: `{ items: T[], total: number, page: number, pageSize: number }`\n",
    "version": "1.0.0",
    "contact": {
      "name": "BG Accountability",
      "url": "https://bridgegaps.app",
      "email": "support@bridgegaps.app"
    },
    "license": {
      "name": "Proprietary",
      "url": ""
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3002",
      "description": "Local Development"
    },
    {
      "url": "https://api.bridgegaps.app",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT access token"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "description": "Minimum 10 characters with uppercase, lowercase, number, and symbol"
          },
          "businessType": {
            "type": "string",
            "enum": [
              "Solopreneur",
              "Startup",
              "MSME"
            ]
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "businessType"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "mfaCode": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Refresh token"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "MfaCodeDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "6-digit TOTP code"
          }
        },
        "required": [
          "code"
        ]
      },
      "AcceptInviteDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "description": "Minimum 10 characters with uppercase, lowercase, number, and symbol"
          }
        },
        "required": [
          "token",
          "password"
        ]
      },
...
      "CreateActivityDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "priority": {
            "type": "string"
          },
          "dueDate": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "default": "Active"
          }
        },
        "required": [
          "title",
          "category"
        ]
      },
      "UpdateActivityDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "priority": {
            "type": "string"
          },
          "dueDate": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
    }
  }
}